Move `layout` under `common`
