+++ src/common/layout/footer/Footer.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Col, Container, Row } from 'reactstrap';
-import Icon from '../../../common/Icon';
-import LocalizedLink from '../../../common/LocalizedLink';
+import Icon from '../../Icon';
+import LocalizedLink from '../../LocalizedLink';
 
 import './footer.scss';
