+++ src/common/layout/koroSection/koroSection.scss
