+++ src/components/pages/winterStoragePage/WinterStoragePage.tsx
@@ -8,8 +8,8 @@ import Hero from '../../../common/hero/Hero';
 import { IconNames } from '../../../common/Icon';
 import Map from '../../../common/Map/Map';
 import UnRegisteredBoatDetails from '../../forms/fragments/UnRegisteredBoatDetails';
-import KoroSection from '../../layout/koroSection/KoroSection';
-import Layout from '../../layout/Layout';
+import KoroSection from '../../../common/layout/koroSection/KoroSection';
+import Layout from '../../../common/layout/Layout';
 import WinterStorageLegend from '../../legends/winterStorageLegend/WinterStorageLegend';
 
 import { BoatTypes } from '../../../types/boatTypes';
